feat(chatbot): implement MongoDB and LLM clients; refactor imports and remove unused response file
