--- backend/chatbot/response_file.txt
@@ -1,11 +1,0 @@
-Anh đang tìm kiếm quà tặng sinh nhật cho bạn gái thanh thiếu niên điều gì đó dễ thương và phù hợp với sở thích của cô nàng. Em có một số gợi ý như sau:
-
-Cắt tóc cho bé gái hình thỏ bông dễ thương (Giá sau giảm: 19000) - Đây là một món quà dễ thương và ý nghĩa cho bạn gái của anh. Hình ảnh của chú thỏ dễ thương chắc chắn sẽ làm cô nàng thích thú.
-
-Hộp nơ kẹp tóc 18 chi tiết cho bé gái siêu dễ thương (Giá: 56000, Giá sau giảm: 56000) - Với 18 chi tiết khác nhau, hộp nơ kẹp tóc này sẽ mang đến cho bạn gái của anh một món quà đa dạng và thú vị.
-
-Mũ đính tóc giả dễ thương cho bé gái (Giá sau giảm: 99000) - Mũ này được làm từ chất liệu cotton co giãn và có tóc giả đính trên mũ, giúp tăng thêm vẻ ngoài dễ thương cho mũ.
-
-Vali kéo du lịch cho trẻ em hình hộp chữ nhật nhiều hình ảnh màu sắc rực rỡ (Giá sau giảm: 299000) - Nếu bạn gái của anh yêu thích đi du lịch, vali này sẽ là một món quà hoàn hảo cho cô nàng. Vali được làm từ nhựa PC và có bánh xe xoay 360 độ, giúp dễ dàng di chuyển.
-
-Em hy vọng những gợi ý này sẽ giúp anh tìm được món quà tặng sinh nhật phù hợp cho bạn gái của mình!
